installer move to bin folder
