+++ filelist.txt
@@ -36,18 +36,19 @@
 /bin/host.lua
 /bin/power.lua
 /bin/fastupdate.lua
-/bin/update.lua
 /bin/warnings.lua
-/bin/programs.lua
-/bin/modLoader.lua
-/bin/connect.lua
+/bin/update.lua
 /bin/splash.lua
+/bin/connect.lua
+/bin/modLoader.lua
 /bin/shortcut.lua
+/bin/installer.lua
+/bin/programs.lua
 /bin/depth.lua
-/bin/components.lua
+/bin/disk.lua
 /bin/eves.lua
 /bin/back.lua
-/bin/disk.lua
+/bin/components.lua
 /lib/event.lua
 /lib/distfs.lua
 /lib/eventTransfer.lua
